use project colour for backgrounds
